add vl53l1x distance mode and timing budget config:
this is the absolutely worst I2C interface and 'use'
of I2C registers i've EVER seen in my entire career
and by a very large margin. whoever designed the timing
budget and distance mode registers was not in this
spacetime we share. VL53L1X is a memorable name
after all

having to set several registers to random magical values
and depending on the distance mode/timing budget instead
of simply writing the desired values to a single register
is mysterious. if this was a hobbyists first time adding
a control mcu to a sensor and interfacing with i2c or
something that'd be a different story. but this is an
expensive, long running STsensor that says state of the
art in the data sheet. it very may well be with its 940nm
invisible laser but it's hard to care when the interface
is this so insanely bad to use. I can only imagine what
kinda of internal cruft and years of legacy ineptitude
is in the way (VL53L0X prob?) from making this register
set not total ass to use.

slightly annoyed but its done. now i can set fast mode :3